requirements.txt
# Core dependencies
numpy==1.26.4
pydantic==2.10.3
python-dotenv==1.1.0
pillow==11.1.0

# Web frameworks
dash==2.17.1
fastapi==0.112.1
httpx==0.28.1

# Configuration
hydra-core==1.3.2
omegaconf>=2.3.0

# LangChain ecosystem
langchain==0.2.16
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langchain-community==0.2.16
langgraph==0.0.69
langchain-google-genai==1.0.8

# Google AI
google-generativeai==0.7.2

# Embeddings and ML
sentence-transformers==3.0.1
huggingface-hub==0.36.0
tokenizers==0.22.1

# PDF processing
PyMuPDF==1.24.10
pytesseract==0.3.10

